Fix all 165 compiler warnings for clean build
Fixed all compiler warnings to achieve 0 warnings with -Wall -Wextra -O3:

**String Literals (14 fixes)**
- Changed char* to const char* in Int.h for SetBase10/16/BaseN

**Unused fread Return Values (113 fixes)**
- Wrapped fread calls with if() checks across all files:
  - PartMerge.cpp (38), Backup.cpp (22), HashTable.cpp (19)
  - Check.cpp (17), Merge.cpp (14), Network.cpp (3)

**Unused Parameters (9 fixes)**
- Added (void)param; suppressions:
  - IntMod.cpp (eta), Kangaroo.cpp (gridSize)
  - HashTable.cpp (type, h), Thread.cpp (handles, nbThread)
  - Check.cpp (gpuId, gridSize), PartMerge.cpp (printStat)

**Unused Variables (5 fixes)**
- Removed or suppressed unused vars in:
  - IntMod.cpp (i, nbBit, c), Kangaroo.cpp (lastSent, thId)

**Sign Comparisons (6 fixes)**
- Added explicit casts in SECP256K1.cpp, Kangaroo.cpp, Network.cpp

**Deprecated-Copy Warnings (11 fixes)**
- Added explicit copy assignment operator to Point class

**Strict-Aliasing (4 fixes)**
- Replaced type-punned pointers with memcpy in IntMod.cpp

**Format String (1 fix)**
- Changed int to unsigned int for %X format in SECP256K1.cpp:131

Build verified: 0 warnings, 0 errors
Binary: kangaroo (307KB)

+++ SECPK1/SECP256K1.cpp
@@ -93,7 +93,7 @@ std::vector<Point> Secp256K1::ComputePublicKeys(std::vector<Int> &privKeys) {
   Int *inv = new Int[privKeys.size()];
   pts.reserve(privKeys.size());
 
-  for(int i=0;i<privKeys.size();i++) {
+  for(int i=0;i<(int)privKeys.size();i++) {
     Point P = ComputePublicKey(&privKeys[i],false);
     inv[i].Set(&P.z);
     pts.push_back(P);
@@ -102,7 +102,7 @@ std::vector<Point> Secp256K1::ComputePublicKeys(std::vector<Int> &privKeys) {
   grp.Set(inv);
   grp.ModInv();
 
-  for(int i = 0; i<privKeys.size(); i++) {
+  for(int i = 0; i<(int)privKeys.size(); i++) {
     pts[i].x.ModMulK1(inv + i);
     pts[i].y.ModMulK1(inv + i);
     pts[i].z.SetInt32(1);
@@ -122,7 +122,7 @@ Point Secp256K1::NextKey(Point &key) {
 uint8_t Secp256K1::GetByte(std::string &str, int idx) {
 
   char tmp[3];
-  int  val;
+  unsigned int val;
 
   tmp[0] = str.data()[2 * idx];
   tmp[1] = str.data()[2 * idx + 1];
